Moved area series into a subfolder.
