+++ js/masters/highcharts.src.js
@@ -23,7 +23,7 @@ import '../Core/Chart/Chart.js';
 import '../Extensions/ScrollablePlotArea.js';
 import '../Extensions/Stacking.js';
 import '../Core/Dynamics.js';
-import '../Series/AreaSeries.js';
+import '../Series/Area/AreaSeries.js';
 import '../Series/SplineSeries.js';
 import '../Series/AreaSplineSeries.js';
 import '../Series/Column/ColumnSeries.js';
